Split engraving into multiple files
