+++ src/engraving.ts
@@ -1,179 +1,12 @@
 import {
   EngravingMask,
   EngravingChisel,
-  ActionError,
-  EngravingActionFunctionResult,
-  OnFinishError,
-  EngravingOnFinishFunctionResult,
 } from './api-model.js';
-import { ActionModel, SingleEngravingModel } from './engraving-model.js';
-import { willFail } from './railway.js';
+import { getLogger, getAlerter } from './chisel-lookup.js';
+import { isFulfilled } from './guards.js';
 import { runValidation, runShield } from './run-validation.js';
-
-const getLogger = (
-  chisel: EngravingChisel,
-  defaultName: string,
-  name?: string
-) => {
-  const func =
-    typeof name === 'string' ? chisel.loggerFunctions[name] : undefined;
-  if (func !== undefined) {
-    return func;
-  }
-  const defaultFunc = chisel.loggerFunctions[defaultName];
-  if (defaultFunc !== undefined) {
-    return defaultFunc;
-  }
-  throw Error(`Neither ${name} not ${defaultName} were available as logger`);
-};
-
-const getAlerter = (
-  chisel: EngravingChisel,
-  defaultName: string,
-  name?: string
-) => {
-  const func =
-    typeof name === 'string' ? chisel.alerterFunctions[name] : undefined;
-  if (func !== undefined) {
-    return func;
-  }
-  const defaultFunc = chisel.alerterFunctions[defaultName];
-  if (defaultFunc !== undefined) {
-    return defaultFunc;
-  }
-  throw Error(`Neither ${name} not ${defaultName} were available as alerter`);
-};
-
-const getUses = (chisel: EngravingChisel, name: string) => {
-  const func =
-    typeof name === 'string' ? chisel.actionFunctions[name] : undefined;
-  if (func !== undefined) {
-    return func;
-  }
-
-  throw Error(`Action uses function ${name} was not available`);
-};
-
-const geOnFinishtUses = (chisel: EngravingChisel, name: string) => {
-  const func =
-    typeof name === 'string' ? chisel.onFinishFunctions[name] : undefined;
-  if (func !== undefined) {
-    return func;
-  }
-
-  throw Error(`OnFinish uses function ${name} was not available`);
-};
-
-const isActionError = (value: unknown): value is ActionError => true;
-
-const createActionError = (
-  name: string,
-  mask: EngravingMask,
-  durationOrderOfMagnitude: number,
-  message: string
-): ActionError => ({
-  txId: mask.txId,
-  engraving: mask.name,
-  durationMagnitude: durationOrderOfMagnitude,
-  action: name,
-  metadata: {},
-  messages: [message],
-});
-
-const createFinishError = (
-  mask: EngravingMask,
-  durationOrderOfMagnitude: number,
-  message: string
-): OnFinishError => ({
-  txId: mask.txId,
-  engraving: mask.name,
-  durationMagnitude: durationOrderOfMagnitude,
-  metadata: {},
-  messages: [message],
-});
-
-const orderOfMagnitude = (started: number, finished: number): number => {
-  const diff = finished - started;
-  if (diff <= 0) return 0;
-  return `${diff}`.length;
-};
-
-const runAction = async (
-  name: string,
-  action: ActionModel,
-  mask: EngravingMask,
-  chisel: EngravingChisel
-): Promise<EngravingActionFunctionResult> => {
-  const started = Date.now();
-  try {
-    const uses = getUses(chisel, action.uses);
-
-    return await uses(mask);
-  } catch (error) {
-    const finished = Date.now();
-    if (isActionError(error)) {
-      return willFail(error);
-    }
-    if (error instanceof Error) {
-      return willFail(
-        createActionError(
-          name,
-          mask,
-          orderOfMagnitude(started, finished),
-          error.message
-        )
-      );
-    }
-
-    return willFail(
-      createActionError(
-        name,
-        mask,
-        orderOfMagnitude(started, finished),
-        '(542921) action default error'
-      )
-    );
-  }
-};
-
-const runOnFinish = async (
-  onFinish: SingleEngravingModel['phases']['onFinish'],
-  actionResults: EngravingActionFunctionResult[],
-  mask: EngravingMask,
-  chisel: EngravingChisel
-): Promise<EngravingOnFinishFunctionResult> => {
-  const started = Date.now();
-  try {
-    const uses = geOnFinishtUses(chisel, onFinish.uses);
-    return await uses({ engravingInput: mask, actionResults });
-  } catch (error) {
-    const finished = Number(Date.now());
-    if (isActionError(error)) {
-      return willFail(error);
-    }
-    if (error instanceof Error) {
-      return willFail(
-        createFinishError(
-          mask,
-          orderOfMagnitude(started, finished),
-          error.message
-        )
-      );
-    }
-
-    return willFail(
-      createFinishError(
-        mask,
-        orderOfMagnitude(started, finished),
-        '(936033) onFinish default error'
-      )
-    );
-  }
-};
-
-const isFulfilled = <T>(
-  input: PromiseSettledResult<T>
-): input is PromiseFulfilledResult<T> => input.status === 'fulfilled';
+import { runAction } from './run-action.js';
+import { runOnFinish } from './run-on-finish.js';
 
 export const runEngraving = async ({
   mask,
