+++ src/utility.ts
@@ -1,0 +1,6 @@
+export const orderOfMagnitude = (started: number, finished: number): number => {
+  const diff = finished - started;
+  if (diff <= 0)
+    return 0;
+  return `${diff}`.length;
+};
